+++ CloudShell.ResourceDefinitions.ReferenceProviders/ContainerApplication/ContainerApplicationResourceTypeProvider.cs
@@ -168,7 +168,7 @@ public sealed class ContainerApplicationResourceTypeProvider :
         {
             if (reference.TypeId is { } typeId &&
                 IsContainerHostResourceType(typeId) &&
-                reference.TryGetResourceId(out containerHostResourceId))
+                reference.TryGetDependsOnResourceId(out containerHostResourceId))
             {
                 return true;
             }