+++ CloudShell.ResourceDefinitions.ReferenceProviders/NameMapping/NameMappingGraphValidator.cs
@@ -30,7 +30,7 @@ public sealed class NameMappingGraphValidator : IResourceDefinitionGraphValidato
 
         foreach (var reference in resource.State.ResourceDependencies)
         {
-            if (!reference.TryGetResourceId(out var resourceId))
+            if (!reference.TryGetDependsOnResourceId(out var resourceId))
             {
                 continue;
             }